requirements.txt
flask==3.0.3
google-genai==1.10.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
lxml==5.2.2
gunicorn==22.0.0
Werkzeug==3.0.6
